Updated for 1.0.6 release

+++ Examples/StackViewExample/StackViewExample/ViewController.m
@@ -32,7 +32,7 @@
 
 - (IBAction)sliderAction:(UISlider *)sender
 {
-    self.stackView.contentSpacing = sender.value;
+    self.stackView.contentSpacing = (CGFloat)sender.value;
 }
 
 @end
